+++ crates/notan_winit/src/gl_manager.rs
@@ -205,7 +205,7 @@ struct InnerSupport {
 }
 
 fn check_support(required_samples: u8, needs_transparency: bool, conf: &GConfig) -> InnerSupport {
-    let req_samples = conf.num_samples() == required_samples;
+    let req_samples = conf.num_samples() == required_samples && required_samples != 0;
     let srgb = conf.srgb_capable();
     let supports_transparency = conf.supports_transparency().unwrap_or(false);
     let transparency = if needs_transparency {