+++ crates/notan_egui/src/lib.rs
@@ -7,4 +7,5 @@ pub use config::EguiConfig;
 pub use extension::{EguiCallbackFn, EguiExtension, EguiRegisterTexture};
 pub use plugin::{EguiPlugin, EguiPluginSugar};
 
+pub use egui::load::SizedTexture;
 pub use egui::*;