+++ CHANGELOG.md
@@ -1,7 +1,14 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## v0.10.0
+## v0.11.0 - 18/10/2023
+
+- Added traits `Serialize` and `Deserialize` to `Color` with the feature `serde` enabled.
+- Updated EGUI to `0.23`.
+- Fixed an error acquiring the GL Context due required samples configuration.
+
+## v0.10.0 - 11/09/2023
+
 - Added `WindowConfig::set_position` to set x/y position before creating the window.
 - Changed `Renderer.begin` uses `Option<ClearOption>` instead of `Option<&ClearOption>`.
 - Changed sizes and positions for Window and Textures from `i32` to `u32`.